Camera parameter를 fovy, aspect로 바꾸었고,  upVector를 추가하였습니다. 플레이어가 움직일 수 있게 하였습니다.(좌우 키로 회전, 상하 키로 이동) Draw 파트를 제외하곤 빌드가 되도록 하였습니다.

+++ GraphicsAssign1/GraphicsAssign1/Objects.h
@@ -136,13 +136,13 @@ public:
 		CollisionComponent_(Transform* collisionObject, Node<ObjectWithComponent*, string>* parentObject);
 		~CollisionComponent_();
 		const Vector3 GetWorldPos() const;
-		inline const int Getxlen() const {
+		inline const GLdouble Getxlen() const {
 			return collisionObject->xlen;
 		}
-		inline const int Getzlen() const {
+		inline const GLdouble Getzlen() const {
 			return collisionObject->zlen;
 		}
-		inline const int Getylen() const {
+		inline const GLdouble Getylen() const {
 			return collisionObject->ylen;
 		}
 		inline const Object::Shape GetShape() const {
